requirements.txt
streamlit==1.22.0
scikit-learn==1.2.2
pandas==2.2.2
numpy==1.26.2
joblib==1.3.2
Pillow==10.0.0
